Move pkg/util/cache to apimachinery
Will be used by client-go as well
